# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "REST API documentation for Tatum Middleware. Tatum Middleware is a client to Tatum Blockchain API Core.",
    "version": "1.0.2",
    "title": "Tatum Middleware API"
  },
  "host": "localhost:6543",
  "basePath": "\/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "\/api\/v1\/erc20\/{id}": {
      "post": {
        "tags": [
          "ERC20 Endpoint"
        ],
        "summary": "Create new ERC20 token",
        "description": "<p>Create new ERC20 token with given supply on Ethereum blockchain. Whole supply of newly created ERC20 token is stored in customer's newly created account. Then it is possible to create new Tatum accounts with ERC20 token name as account's currency.</p>",
        "operationId": "createErc20Currency",
        "consumes": [
          "application\/json"
        ],
        "produces": [
          "application\/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "erc20",
            "description": "erc20",
            "required": true,
            "schema": {
              "$ref": "#\/definitions\/Erc20"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Customer internal ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#\/definitions\/Erc20Response"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "400": {
            "description": "Invalid",
            "schema": {
              "$ref": "#\/definitions\/Error"
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal error",
            "schema": {
              "$ref": "#\/definitions\/Error"
            }
          }
        },
        "security": [
          {
            "apiKey": [
            ]
          }
        ],
        "deprecated": false
      }
    },
    "\/api\/v1\/erc20\/{symbol}\/{address}": {
      "post": {
        "tags": [
          "ERC20 Endpoint"
        ],
        "summary": "Set ERC20 token contract address",
        "description": "<p>Set contract address of ERC20 token. This must be done in order to communicate with ERC20 smart contract. After creating and deploying ERC20 token to Ethereum blockchain, smart contract address is generated and must be set within Tatum. Otherwise Tatum platform will not be able to detect incoming deposits of ERC20 and do withdrawals from Tatum accounts to other blockchain addresses.</p>",
        "operationId": "storeErc20ContractAddress",
        "consumes": [
          "application\/json"
        ],
        "produces": [
          "*\/*"
        ],
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "ERC20 contract address",
            "required": true,
            "type": "string"
          },
          {
            "name": "symbol",
            "in": "path",
            "description": "ERC20 symbol name.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "400": {
            "description": "Invalid",
            "schema": {
              "$ref": "#\/definitions\/Error"
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal error",
            "schema": {
              "$ref": "#\/definitions\/Error"
            }
          }
        },
        "security": [
          {
            "apiKey": [
            ]
          }
        ],
        "deprecated": false
      }
    },
...
    "\/api\/v1\/withdrawal\/{id}\/{txId}": {
      "put": {
        "tags": [
          "Withdrawal Endpoint"
        ],
        "summary": "Complete withdrawal",
        "description": "<p>Invoke complete withdrawal as soon as blockchain transaction ID is available. Otherwise withdrawal will be processed automatically in next block and all other withdrawals will be pending unless the last one is processed and marked as compelted..</p>",
        "operationId": "completeWithdrawal",
        "consumes": [
          "application\/json"
        ],
        "produces": [
          "*\/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of created withdrawal",
            "required": true,
            "type": "string"
          },
          {
            "name": "txId",
            "in": "path",
            "description": "Blockchain transaction ID of created withdrawal",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "400": {
            "description": "Invalid",
            "schema": {
              "$ref": "#\/definitions\/Error"
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal error",
            "schema": {
              "$ref": "#\/definitions\/Error"
            }
          }
        },
        "security": [
          {
            "apiKey": [
            ]
          }
        ],
        "deprecated": false
      }
    },
...
    "\/util\/v1\/eth\/wallet": {
      "post": {
        "parameters": [
          {
            "name": "chain",
            "in": "body",
            "description": "chain - 'mainnet' or 'ropsten'",
            "required": true,
            "schema": {
              "$ref": "#\/definitions\/WalletGenerateEth"
            }
          }
        ],
        "operationId": "generateEthWallet",
        "summary": "Generate ETH wallet.",
        "description": "<p>Generate wallet. This will create BIP44 HD wallet with predefined derivation path and returns mnemonic phrase and extended public and private key to be used for address generation.</p>",
        "tags": [
          "Ethereum blockchain helper Endpoint"
        ],
        "responses": {
          "200": {
            "description": "Object containing mnemonic, xpriv and xpub for generated wallet.",
            "schema": {
              "$ref": "#\/definitions\/Wallet"
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error",
            "schema": {
              "$ref": "#\/definitions\/Error"
            }
          }
        }
      }
    },
    "\/util\/v1\/eth\/wallet\/xpub\/{pub}\/{i}": {
      "get": {
        "parameters": [
          {
            "name": "pub",
            "in": "path",
            "description": "xpub to generate address from",
            "required": true,
            "type": "string"
          },
          {
            "name": "i",
            "in": "path",
            "description": "derivation index of address",
            "required": true,
            "type": "integer"
          }
        ],
        "operationId": "calculateEthAddress",
        "summary": "Calculate ETH address",
        "description": "<p>Calculate address from xpub on Ropsten \/ Mainnet for given derivation path index.</p>",
        "tags": [
          "Ethereum blockchain helper Endpoint"
        ],
        "responses": {
          "200": {
            "description": "Generated address",
            "schema": {
              "$ref": "#\/definitions\/GeneratedAddress"
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error",
            "schema": {
              "$ref": "#\/definitions\/Error"
            }
          }
        }
      }
    },
    "\/util\/v1\/eth\/wallet\/xpriv": {
      "post": {
        "parameters": [
          {
            "name": "xpriv",
            "in": "body",
            "description": null,
            "required": true,
            "schema": {
              "$ref": "#\/definitions\/XPrivEth"
            }
          }
        ],
        "operationId": "calculateEthPrivKey",
        "summary": "Calculate private key of ETH address",
        "description": "<p>Calculate private key of address from mnemonic on Ropsten \/ Mainnet for given derivation path index.</p>",
        "tags": [
          "Ethereum blockchain helper Endpoint"
        ],
        "responses": {
          "200": {
            "description": "Generated private key",
            "schema": {
              "$ref": "#\/definitions\/PrivKey"
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error",
            "schema": {
              "$ref": "#\/definitions\/Error"
            }
          }
        }
      }
    },
    "\/util\/v1\/eth\/transfer": {
      "post": {
        "parameters": [
          {
            "name": "transfer",
            "in": "body",
            "description": null,
            "required": true,
            "schema": {
              "$ref": "#\/definitions\/EthTransfer"
            }
          }
        ],
        "operationId": "transferEth",
        "summary": "Send ETH \/ Ropsten ETH from account to account",
        "description": "Send ETH \/ Ropsten ETH from account to account. This will create Tatum internal withdrawal request with ID. If every system works as expected, withdrawal request is marked as complete and transaction id is assigned to it.<ul><li>If Ethereum server connection is unavailable, withdrawal request is cancelled.</li><li>If blockchain transfer is successful, but is it not possible to reach Tatum Blockchain API, transaction id of blockchain transaction is returned and withdrawal request must be completed manually, otherwise all other withdrawals will be pending.</li></ul></p>",
        "tags": [
          "Ethereum blockchain helper Endpoint"
        ],
        "responses": {
          "200": {
            "description": "txHash of successful transaction",
            "schema": {
              "$ref": "#\/definitions\/TxHash"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "400": {
            "description": "Invalid",
            "schema": {
              "$ref": "#\/definitions\/Error"
            }
          },
          "500": {
            "description": "Internal error",
            "schema": {
              "$ref": "#\/definitions\/Error"
            }
          }
        },
        "security": [
          {
            "apiKey": [
            ]
          }
        ]
      }
    },
    "\/util\/v1\/eth\/erc20\/deploy": {
      "post": {
        "parameters": [
          {
            "name": "erc20deploy",
            "in": "body",
            "description": null,
            "required": true,
            "schema": {
              "$ref": "#\/definitions\/Erc20Deploy"
            }
          }
        ],
        "operationId": "deployErc20",
        "summary": "Deploy ETH \/ Ropsten ETH ERC20 Smart Contract.",
        "description": "<p>Deploy ETH \/ Ropsten ETH ERC20 Smart Contract. Response could take quite a lot of time, average time of creation is 3-4 minutes. After deploying contract to blockchain, contract address will become available and must be stored within Tatum, otherwise it will not be possible to interact with it.</p>",
        "tags": [
          "Ethereum blockchain helper Endpoint"
        ],
        "responses": {
          "200": {
            "description": "information about ERC20 smart contract",
            "schema": {
              "$ref": "#\/definitions\/Erc20Token"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "400": {
            "description": "Invalid",
            "schema": {
              "$ref": "#\/definitions\/Error"
            }
          },
          "500": {
            "description": "Internal error",
            "schema": {
              "$ref": "#\/definitions\/Error"
            }
          }
        },
        "security": [
          {
            "apiKey": [
            ]
          }
        ]
      }
    },
    "\/util\/v1\/eth\/erc20\/transfer": {
      "post": {
        "parameters": [
          {
            "name": "erc20",
            "in": "body",
            "description": null,
            "required": true,
            "schema": {
              "$ref": "#\/definitions\/Erc20Transfer"
            }
          }
        ],
        "operationId": "transferErc20",
        "summary": "Transfer ETH \/ Ropsten ETH ERC20",
        "description": "<p>Transfer ETH \/ Ropsten ETH ERC20 Smart Contract Tokens from account to account. This will create Tatum internal withdrawal request with ID. If every system works as expected, withdrawal request is marked as complete and transaction id is assigned to it.<ul><li>If Ethereum server connection is unavailable, withdrawal request is cancelled.</li><li>If blockchain transfer is successful, but is it not possible to reach Tatum Blockchain API, transaction id of blockchain transaction is returned and withdrawal request must be completed manually, otherwise all other withdrawals will be pending.</li></ul></p>",
        "tags": [
          "Ethereum blockchain helper Endpoint"
        ],
        "responses": {
          "200": {
            "description": "txHash of successful transaction",
            "schema": {
              "$ref": "#\/definitions\/TxHash"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "400": {
            "description": "Invalid",
            "schema": {
              "$ref": "#\/definitions\/Error"
            }
          },
          "500": {
            "description": "Internal error",
            "schema": {
              "$ref": "#\/definitions\/Error"
            }
          }
        },
        "security": [
          {
            "apiKey": [
            ]
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "apiKey": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header",
      "description": "JWT token generated from `apiKey` field of your API Key and current timestamp signed by `secret` of your API Key passed as Bearer token in authorization header, e.g. `authorization: Bearer 12kjn128n123n12i3n`.<br/>`{\"apiKey\": \"your_api_key\", \"created\": current_timestamp_millis}`<br/>How to create valid **JWT** can be found at <a href=\"https://jwt.io\">https://jwt.io</a>, or you can use our JWT helper Endpoint."
    }
  },
...
  "produces": [
    "application\/json"
  ],
  "consumes": [
    "application\/json"
  ]
}